+++ astropy/table/tests/test_column.py
@@ -764,3 +764,21 @@ def test_unicode_sandwich_set(Column):
     c[:] = ''
     c[:] = [uba, b'def']
     assert np.all(c == [uba, b'def'])
+
+
+@pytest.mark.parametrize('class1', [table.MaskedColumn, table.Column])
+@pytest.mark.parametrize('class2', [table.MaskedColumn, table.Column, str, list])
+def test_unicode_sandwich_compare(class1, class2):
+    """Test that comparing a bytestring Column/MaskedColumn with various
+    str (unicode) object types gives the expected result.  Tests #6838.
+    """
+    obj1 = class1([b'a', b'b'])
+    if class2 is str:
+        obj2 = 'a'
+    elif class2 is list:
+        obj2 = ['a', 'c']
+    else:
+        obj2 = class2(['a', 'c'])
+
+    assert np.all((obj1 == obj2) == [True, False])
+    assert np.all((obj2 == obj1) == [True, False])